+++ R/summary.colspace.R
@@ -55,6 +55,11 @@
 #'  as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.
 
 summary.colspace <- function(object, by = NULL, ...){
+  
+  if(is.null(attr(object, 'clrsp'))){
+    message('Cannot return full colspace summary on subset data')
+    return(summary(as.data.frame(object))) 
+  }
 
   cat("Colorspace & visual model options:\n",
       '* Colorspace:', attr(object, 'clrsp'), '\n',